refactor generation flow
