+++ server/src/modules/generation/generation.error.ts
@@ -25,7 +25,7 @@ export const normalizeGenerationError = (error: unknown, options?: { path?: stri
         }
     }
 
-    if (normalizedMessage.includes('prompt agent')) {
+    if (normalizedMessage.includes('context agent')) {
         return {
             internalMessage,
             publicMessage:
@@ -52,22 +52,6 @@ export const normalizeGenerationError = (error: unknown, options?: { path?: stri
             publicMessage: options?.path
                 ? `I hit an issue while generating ${getFileLabel(options.path)}. Please retry the build.`
                 : 'I started the build but hit an issue while generating the project files. Please retry the build.',
-        }
-    }
-
-    if (normalizedMessage.includes('edit agent')) {
-        return {
-            internalMessage,
-            publicMessage:
-                'I hit an issue while applying that edit. Please try again with a more specific change request.',
-        }
-    }
-
-    if (normalizedMessage.includes('fix agent')) {
-        return {
-            internalMessage,
-            publicMessage:
-                'I found the preview error but could not repair it automatically. Please try again or adjust the request.',
         }
     }
