+++ server/src/modules/generation/generation.service.ts
@@ -1,11 +1,5 @@
-﻿import {
-    applyProjectEditWorkflow,
-    applyProjectFixWorkflow,
-    generateWebsite,
-} from './generation.workflows'
+﻿import { generateWebsite } from './generation.workflows'
 
 export const generateService = {
     generateWebsite: generateWebsite,
-    applyProjectEdit: applyProjectEditWorkflow,
-    applyProjectFix: applyProjectFixWorkflow,
 }
