+++ server/src/modules/generation/generation.utils.ts
@@ -1,3 +1,10 @@
+export const cleanPrompt = (input: string): string => {
+    if (!input) {
+        return ''
+    }
+    return input.replace(/\r\n?/g, ' ').replace(/\t/g, ' ').replace(/\s+/g, ' ').trim()
+}
+
 const ROOT_FRONTEND_FILE_PATTERNS = [
     /^package\.json$/,
     /^index\.html$/,
