+++ server/src/modules/generation/generation.workflows.ts
@@ -7,7 +7,7 @@ import {
 } from '../../core/engine/generation'
 import { saveProjectFiles } from '../../lib/save-project-files'
 import { prisma } from '../../config/db'
-import { cleanPrompt } from '../../utils/cleanPrompt'
+import { cleanPrompt } from '../generation/generation.utils'
 import { persistCanvasDocument } from '../canvas/canvas.persistence'
 
 import { planAgentResponseSchema, promptAgentResponseSchema } from './generation.schema'
@@ -29,13 +29,7 @@ import {
     publishFinalPreviewSnapshot,
     publishIncrementalPreviewSnapshot,
 } from './generation.runtime'
-import type {
-    ApplyProjectEditInput,
-    ApplyProjectFixInput,
-    GenerateWebsiteInput,
-    ProjectPlan,
-    ProjectRecord,
-} from './generation.types'
+import type { GenerateWebsiteInput, ProjectPlan, ProjectRecord } from './generation.types'
 
 export const generateWebsite = async (data: GenerateWebsiteInput) => {
     const { prompt, onEvent } = data
@@ -346,87 +340,5 @@ export const generateWebsite = async (data: GenerateWebsiteInput) => {
         }
 
         throw error
-    }
-}
-
-export const applyProjectEditWorkflow = async (data: ApplyProjectEditInput) => {
-    const base = await getProjectRevisionBase(data)
-    const prompt = cleanPrompt(data.prompt)
-    const editResult = await applyProjectEdit({
-        prompt,
-        ...(data.selectedElement ? { selectedElement: data.selectedElement } : {}),
-        project: {
-            name: base.project.name,
-            description: base.project.description,
-            prompt: base.project.prompt,
-        },
-        recentMessages: toRecentMessages(base.baseVersion),
-        files: base.baseFiles,
-    })
-
-    const { mergedFiles, appliedFiles, removedFiles } = mergeProjectFiles({
-        currentFiles: base.baseFiles,
-        updatedFiles: editResult.updatedFiles,
-        deletedFiles: editResult.deletedFiles,
-    })
-
-    const persisted = await persistProjectRevision({
-        project: base.project,
-        userId: data.userId,
-        baseVersion: base.baseVersion,
-        nextVersionNumber: base.nextVersionNumber,
-        mergedFiles,
-        removedFiles,
-        sourcePrompt: prompt,
-        assistantMessage: editResult.message,
-        summary: editResult.summary,
-        nextProjectPrompt: prompt,
-        canvasState: data.canvasState,
-    })
-
-    return {
-        ...persisted,
-        appliedFiles,
-        deletedFiles: removedFiles,
-    }
-}
-
-export const applyProjectFixWorkflow = async (data: ApplyProjectFixInput) => {
-    const base = await getProjectRevisionBase(data)
-    const errorMessage = data.errorMessage.trim()
-    const fixResult = await applyProjectFix({
-        errorMessage,
-        ...(data.stack ? { stack: data.stack } : {}),
-        project: {
-            name: base.project.name,
-            description: base.project.description,
-            prompt: base.project.prompt,
-        },
-        recentMessages: toRecentMessages(base.baseVersion),
-        files: base.baseFiles,
-    })
-
-    const { mergedFiles, appliedFiles, removedFiles } = mergeProjectFiles({
-        currentFiles: base.baseFiles,
-        updatedFiles: fixResult.updatedFiles,
-        deletedFiles: fixResult.deletedFiles,
-    })
-
-    const persisted = await persistProjectRevision({
-        project: base.project,
-        userId: data.userId,
-        baseVersion: base.baseVersion,
-        nextVersionNumber: base.nextVersionNumber,
-        mergedFiles,
-        removedFiles,
-        sourcePrompt: `Fix preview error: ${errorMessage}`,
-        assistantMessage: fixResult.message,
-        summary: fixResult.summary,
-    })
-
-    return {
-        ...persisted,
-        appliedFiles,
-        deletedFiles: removedFiles,
     }
 }
