--- server/src/utils/cleanPrompt.ts
@@ -1,6 +1,0 @@
-export const cleanPrompt = (input: string): string => {
-    if (!input) {
-        return ''
-    }
-    return input.replace(/\r\n?/g, ' ').replace(/\t/g, ' ').replace(/\s+/g, ' ').trim()
-}
